% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteration.R
\name{assess_alteration}
\alias{assess_alteration}
\title{Assess hydrologic alteration by flow metric}
\usage{
assess_alteration(percentiles, predictions, ffc_values, comid)
}
\description{
This will be be the function that returns the alteration values for a single segment - it will
return a record for every metric with its alteration assessment.
}
