% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{attach_water_year_data}
\alias{attach_water_year_data}
\title{Add calendar_year/calendar_month/calendar_day/water_year fields}
\usage{
attach_water_year_data(timeseries, date_field = "posix_time")
}
\arguments{
\item{timeseries.}{A timeseries data frame with a date field}

\item{date_field.}{A sortable field of date/time information - a good choice would be a field
created by strptime or any other POSIX format time.}
}
\description{
Attaches fields for the year, month, day and water year to a data frame with a POSIX time field.
Fields will be named calendar_year, calendar_month, calendar_day, and water_year.
}
