% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{clean_account}
\alias{clean_account}
\title{Clean account}
\usage{
clean_account(token)
}
\arguments{
\item{token.}{A valid token for the account to clean out. See install
instructions for details}
}
\description{
Deletes all runs in the online FFC for the user. Helpful in cases
where the website configuration changes and makes accounts broken
in the online interface.
}
