% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteration.R
\name{determine_status}
\alias{determine_status}
\title{Calculate the alteration status of a flow metric}
\usage{
determine_status(
  median,
  predictions,
  assessed_observations,
  metric,
  days_in_water_year,
  annual
)
}
\arguments{
\item{median}{The calculated median value from the observed data}

\item{predictions}{The predicted metric values for this specific metric - should have p10, p25, p50, p75, p90 values}

\item{assessed_observations}{vector of raw observed metric values (FFC output) that has already been assessed for whether it is in range
so that records that are low/early are -1, records that are in range are 0, and records that are high/late are 1}

\item{metric}{character name of the metric - case sensitive. Currently only used for timing metrics, which must have "_Tim" in the name}

\item{days_in_water_year}{numeric of how many days in the water year (typically 365, but could be 366).}
}
\description{
This method returns an alteration status record for a specific flow metric, but requires the calculated FFC percentiles,
a lower and upper bound, and a set of observations that have already been assessed for whether they're within that lower
or upper bound so that they are -1 for low/early, 0 for within range, and 1 for high/late. They need to already be assessed
because some metrics (*ahem* timing) need their own ways to assess low/high, or early/late
}
