% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\docType{package}
\name{ffcAPIClient}
\alias{ffcAPIClient}
\title{ffcAPIClient: Processes time-series flow data using the online functional
flows calculator}
\description{
For now, see the documentation for \code{\link{get_ffc_results_for_df}}
}
\examples{
# Initialize a Run
test_data <- example_gagedata()  # just get some fake gage data - based on Daniel Philippus' code - you can build your own data frame here
ffcAPIClient::set_token(YOUR_TOKEN_VALUE_IN_QUOTES)  # you'll need to get your own of this - see above
results <- ffcAPIClient::get_ffc_results_for_df(test_data)  # send it to the FFC online to process

# Retrieve Results and Plot
## get the DRH data as a data frame with percentiles for columns and days for rows
drh_data <- ffcAPIClient::get_drh(results)
plot(drh_data$seventy_five, type="l")  # plot the seventy-fifth percentile DRH

}
