% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_metrics}
\alias{flow_metrics}
\title{Modeled flow metric predictions for all stream segments}
\format{A data frame with  rows and  variables:
\describe{
  \item{name}{text}
  \item{name}{text}
  ...
}}
\source{
\url{}
}
\usage{
flow_metrics
}
\description{
Modeled flow metric predictions for all stream segments
}
\keyword{datasets}
