% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{force_consistent_naming}
\alias{force_consistent_naming}
\title{Set Preference to Rename Metrics Consistently}
\usage{
force_consistent_naming(force_to)
}
\arguments{
\item{force_to}{boolean Set to TRUE to enable metric renaming. Set to FALSE to disable it.}
}
\description{
The Peak Magnitude flow metrics are named Peak_2, Peak_5, and Peak_10, while other flow metrics
for magnitude use a Component_Mag format. Setting this preference forces all peak metrics
to use the naming Peak_Mag_2, Peak_Mag_5, and Peak_Mag_10. This is inconsistent with
CEFF, but internally consistent with metric names. The default behavior is not to do this.
}
\details{
To use, run force_consistent_naming(TRUE). Then, any function that returns metrics from a
FFCProcessor object (includes the FFCProcessor itself and all evaluate_alteration functions - in the
future should be any exported function) will rename the peak metrics.
To turn off, run force_consistent_naming(FALSE).
}
