% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flow_metrics.R
\name{get_comid_for_lon_lat}
\alias{get_comid_for_lon_lat}
\title{Retrieves COMID for a given USGS gage which collects daily data.}
\usage{
get_comid_for_lon_lat(longitude, latitude)
}
\arguments{
\item{longitude}{numeric. Longitude or X.}

\item{latitude}{numeric. Longitude or Y.}
}
\description{
This function returns the COMID associated with a specific USGS gage.
It can be used to associate gage data with flow metric predictions a
stream segment identified with the \code{com_id} input variable.
}
