% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{get_drh}
\alias{get_drh}
\title{Returns the dimensionless reference hydrograph results as a data frame}
\usage{
get_drh(results)
}
\description{
Returns the dimensionless reference hydrograph results as a data frame
}
