% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_class.R
\name{get_ffc_parameters_for_comid}
\alias{get_ffc_parameters_for_comid}
\title{Get the parameters sent to the FFC for a stream segment}
\usage{
get_ffc_parameters_for_comid(comid)
}
\arguments{
\item{comid}{An NHD stream segment COMID}
}
\description{
Given a COMID, looks up the hydrogeomorphic stream classification, then uses that to find the default parameters that
should be sent to the FFC online for that stream class. Returns a nested list of parameters to send to the FFC.
}
