% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{get_ffc_results_for_usgs_gage}
\alias{get_ffc_results_for_usgs_gage}
\title{Run Gage Data Through the Functional Flows Calculator}
\usage{
get_ffc_results_for_usgs_gage(gage_id)
}
\arguments{
\item{gage_id}{integer. The USGS Gage ID value for the gage you want to return timeseries data for}
}
\value{
list. Functional Flow Calculator results
}
\description{
Provided with an integer Gage ID, this function pulls the timeseries data for the
gage and processes it in a single step. Returns the functional flow calculator's results list.
}
