% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{get_results_as_df}
\alias{get_results_as_df}
\title{Convert FFC results list to data frame with metric names}
\usage{
get_results_as_df(results, drop_fields)
}
\description{
More documentation forthcoming
}
