% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{get_results_for_name}
\alias{get_results_for_name}
\title{Retrieve processed results from FFC.}
\usage{
get_results_for_name(name, autodelete)
}
\arguments{
\item{name}{the name of the run to retrieve from the online FFC}

\item{autodelete}{when TRUE, deletes the run in the online FFC, if found. When FALSE, leaves run in FFC online for later
retrieval.}
}
\description{
Gets the results for the given named run of the FFC. Returns the nested list - all other processing must be handled
by the caller.
}
