% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{get_token}
\alias{get_token}
\title{Retrieves a token previously set by set_token in the same R session}
\usage{
get_token()
}
\description{
Retrieves a token previously set by set_token in the same R session
}
