% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gages.R
\name{get_usgs_gage_data}
\alias{get_usgs_gage_data}
\title{Retrieves USGS timeseries gage data}
\usage{
get_usgs_gage_data(gage_id)
}
\arguments{
\item{gage_id}{integer. The USGS Gage ID value for the gage you want to return timeseries data for}
}
\value{
dataframe. Will include a flow field (CFS) and a date field (MM/DD/YYYY)
}
\description{
This is just a helper function that calls the gage constructor, gets the flows and returns them in one step.
Useful in situations where we don't need the flexibility of the USGSGage class
}
