% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{make_flow_json}
\alias{make_flow_json}
\title{Makes the part of the JSON string that is just for the flow data - needs to be passed
into make_json later as "data_json"}
\usage{
make_flow_json(flows_df, flow_field, date_field)
}
\description{
Makes the part of the JSON string that is just for the flow data - needs to be passed
into make_json later as "data_json"
}
