% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{make_json}
\alias{make_json}
\title{Prepares the JSON payload to send to the eflows website.}
\usage{
make_json(data_json, start_date, token, extra)
}
\description{
Prepares the JSON payload to send to the eflows website.
}
