% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{merge_list}
\alias{merge_list}
\title{Merges Data Frames by Year Column}
\usage{
merge_list(df1, df2)
}
\description{
Just a simple function that can be used with Reduce to merge multiple data frames together by year
}
