% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{plot_drh}
\alias{plot_drh}
\title{Plots the Dimensionless Reference Hydrograph}
\usage{
plot_drh(results, output_path = NULL)
}
\arguments{
\item{results}{list.}

\item{output_path, }{default NULL. Optional. When set, saves the DRH plot to the output
file path provided.}
}
\description{
Given a set of results data from get_ffc_results_for_df or get_ffc_results_for_usgs_gage,
processes the DRH data and returns a plot object.
}
\details{
Credit to Ryan Peek for the plotting code in this function.
}
