% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{process_data}
\alias{process_data}
\title{Sends flow timeseries data off to the functional flows calculator. Does *not* retrieve results!}
\usage{
process_data(flows_df, flow_field, date_field, start_date, name)
}
\description{
Sends flow timeseries data off to the functional flows calculator. Does *not* retrieve results!
}
