% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stream_class_data}
\alias{stream_class_data}
\title{Geomorphic Stream Classification}
\format{
A data frame :
\describe{
  \item{CLASS}{The stream classification ID}
  \item{COMID}{The NHD COMID of the stream segment}
  \item{CLASS_CODE}{The character stream classification ID - follows the form:
    SM = Snowmelt, HSR = High Volume Snowmelt and Rain, LSR = Low Volume Snowmelt and Rain,
    WS = Winter Storms, GW = Groundwater, PGR = Perennial Groundwater and Rain,
    FER = Flashy Ephemeral Rain, RGW = Rain and Seasonal Groundwater,
    HLP = High elevation, low precipitation
  }
}
\url{https://doi.org/10.1007/s00267-018-1077-7}
}
\usage{
stream_class_data
}
\description{
Contains the geomorphic classification by stream COMID for ~70,000 stream
segments in California (low-order streams excluded). Streams were classified
as described in Lane, Belize A., Samuel Sandoval-Solis, Eric D. Stein,
Sarah M. Yarnell, Gregory B. Pasternack, and Helen E. Dahlke. 2018.
“Beyond Metrics? The Role of Hydrologic Baseline Archetypes in Environmental
 Water Management.” Environmental Management 62 (4): 678–93.
 https://doi.org/10.1007/s00267-018-1077-7.
}
\keyword{datasets}
