% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{timing_alteration}
\alias{timing_alteration}
\title{So here's a pain in the rear - for timing metrics, none of them have percentiles that cross water years
(which seems kind of suspicious to me, but whatever), so the upper bound will never be earlier
in the water year than the lower bound - that makes things a bit easier. BUT, for alteration,
we have plenty of timing metrics that are predicted to be very early in the water year. If the actual
value is early enough that it's in the previous water year (looking at you fall flushing flow), then
we don't want to mark it as being *late* when it's actually early! So, we need to have some rules for early
and late for timing. Planning to determine the range of values that aren't in the inter-80th percentile range
and then find the day of the water year that's in the middle. Timings earlier than that are late, timings after that
are early.}
\usage{
timing_alteration(median_value, low_bound, upper_bound, days_in_water_year)
}
\description{
So here's a pain in the rear - for timing metrics, none of them have percentiles that cross water years
(which seems kind of suspicious to me, but whatever), so the upper bound will never be earlier
in the water year than the lower bound - that makes things a bit easier. BUT, for alteration,
we have plenty of timing metrics that are predicted to be very early in the water year. If the actual
value is early enough that it's in the previous water year (looking at you fall flushing flow), then
we don't want to mark it as being *late* when it's actually early! So, we need to have some rules for early
and late for timing. Planning to determine the range of values that aren't in the inter-80th percentile range
and then find the day of the water year that's in the middle. Timings earlier than that are late, timings after that
are early.
}
