% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{evaluate_alteration}
\alias{evaluate_alteration}
\title{Generate FFC Results and Plots for Timeseries Data}
\usage{
evaluate_alteration(
  timeseries_df,
  token,
  comid,
  longitude,
  latitude,
  plot_output_folder,
  date_format_string
)
}
\description{
Generate FFC Results and Plots for Timeseries Data
}
