% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{evaluate_gage_alteration}
\alias{evaluate_gage_alteration}
\title{Generate FFC Results and Plots for Gage Data}
\usage{
evaluate_gage_alteration(gage_id, token, plot_output_folder)
}
\arguments{
\item{plot_output_folder}{}
}
\description{
Generate FFC Results and Plots for Gage Data
}
