% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\docType{package}
\name{ffcAPIClient}
\alias{ffcAPIClient}
\title{ffcAPIClient: Processes time-series flow data using the online functional
flows calculator}
\description{
For now, see the documentation for \code{\link{evaluate_alteration}} and
\code{\link{evaluate_gage_alteration}}
}
\examples{
\dontrun{
# If you have a gage and a token, you can get all results simply by running
ffcAPIClient::evaluate_gage_alteration(gage_id = 11427000, token = "your_token", plot_output_folder = "C:/Users/youruser/Documents/NFA_Gage_Alteration")
# output_folder is optional. When provided, it will save plots there. It will show plots regardless.

# If you have a data frame with flow and date fields that isn't a gage, you can run
ffcAPIClient::evaluate_alteration(timeseries_df = your_df, token = "your_token", plot_output_folder = "C:/Users/youruser/Documents/Timeseries_Alteration", comid=yoursegmentcomid)
# it also *REQUIRES* you provide either a comid argument with the stream segment COMID, or both
# longitude and latitude arguments.
# Make sure that dates are in the same format as the FFC requires on its website. We may add reformatting in the future

}
}
