% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{FFCProcessor}
\alias{FFCProcessor}
\title{FFCProcessor Class}
\description{
The new workhorse of the client - this class is meant to bring together the scattershot functions
in other parts of the package so that data can be integrated into a single class with a single
set of tasks. Other functions are likely to be supported for a while (and this may even rely on them),
but long run, much of the code in this file might move into this class, with the shortcut functions
creating this class behind the scenes and returning an instance of this object.
}
\details{
More details to come, and more examples. For now, still use the general functions \code{\link{evaluate_alteration}}
and \code{\link{evaluate_gage_alteration}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set_up}{\code{FFCProcessor$set_up()}}
\item \href{#method-run}{\code{FFCProcessor$run()}}
\item \href{#method-rename_inconsistent_metrics}{\code{FFCProcessor$rename_inconsistent_metrics()}}
\item \href{#method-get_gage_id}{\code{FFCProcessor$get_gage_id()}}
\item \href{#method-step_one_functional_flow_results}{\code{FFCProcessor$step_one_functional_flow_results()}}
\item \href{#method-step_two_explore_ecological_flow_criteria}{\code{FFCProcessor$step_two_explore_ecological_flow_criteria()}}
\item \href{#method-step_three_assess_alteration}{\code{FFCProcessor$step_three_assess_alteration()}}
\item \href{#method-get_ffc_results}{\code{FFCProcessor$get_ffc_results()}}
\item \href{#method-evaluate_alteration}{\code{FFCProcessor$evaluate_alteration()}}
\item \href{#method-clone}{\code{FFCProcessor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_up"></a>}}
\if{latex}{\out{\hypertarget{method-set_up}{}}}
\subsection{Method \code{set_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FFCProcessor$set_up(gage_id, timeseries, comid, token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FFCProcessor$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rename_inconsistent_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-rename_inconsistent_metrics}{}}}
\subsection{Method \code{rename_inconsistent_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FFCProcessor$rename_inconsistent_metrics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_gage_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_gage_id}{}}}
\subsection{Method \code{get_gage_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FFCProcessor$get_gage_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-step_one_functional_flow_results"></a>}}
\if{latex}{\out{\hypertarget{method-step_one_functional_flow_results}{}}}
\subsection{Method \code{step_one_functional_flow_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FFCProcessor$step_one_functional_flow_results(
  gage_id,
  timeseries,
  comid,
  token,
  output_folder
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-step_two_explore_ecological_flow_criteria"></a>}}
\if{latex}{\out{\hypertarget{method-step_two_explore_ecological_flow_criteria}{}}}
\subsection{Method \code{step_two_explore_ecological_flow_criteria()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FFCProcessor$step_two_explore_ecological_flow_criteria()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-step_three_assess_alteration"></a>}}
\if{latex}{\out{\hypertarget{method-step_three_assess_alteration}{}}}
\subsection{Method \code{step_three_assess_alteration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FFCProcessor$step_three_assess_alteration()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ffc_results"></a>}}
\if{latex}{\out{\hypertarget{method-get_ffc_results}{}}}
\subsection{Method \code{get_ffc_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FFCProcessor$get_ffc_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evaluate_alteration"></a>}}
\if{latex}{\out{\hypertarget{method-evaluate_alteration}{}}}
\subsection{Method \code{evaluate_alteration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FFCProcessor$evaluate_alteration()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FFCProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
