% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flow_metrics.R
\name{get_comid_for_lon_lat}
\alias{get_comid_for_lon_lat}
\title{Retrieves COMID for a given USGS gage which collects daily data.}
\usage{
get_comid_for_lon_lat(longitude, latitude, online = TRUE)
}
\arguments{
\item{longitude}{numeric. Longitude or X.}

\item{latitude}{numeric. Longitude or Y.}

\item{online}{boolean. Default TRUE. When TRUE, looks up the COMID using the nhdplustools
package and USGS web services. Sometimes these are spotty though, so you can set offline
to FALSE and this function will perform the lookup locally with spatial data. It will still
download a large amount of spatial data for
NHD segments the first time it runs with the flag set to FALSE, but then future lookups will be
much faster. It uses the nhdR package, which is not included as a package requirement, and instead
is installed only if you set \code{ffc$get_comid_online = FALSE}.}
}
\description{
This function returns the COMID associated with a specific USGS gage.
It can be used to associate gage data with flow metric predictions a
stream segment identified with the \code{comid} input variable.
}
