% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{get_ffc_results_for_df}
\alias{get_ffc_results_for_df}
\title{This is the primary function to use from the API client itself to obtain
raw FFC results. It will generate a unique ID, run the data frame through
the FFC, and then delete the results for that ID from the website so as not
to clutter up the user's account, or store too much data on the server side.}
\usage{
get_ffc_results_for_df(df, flow_field, date_field, start_date)
}
\description{
This is the primary function to use from the API client itself to obtain
raw FFC results. It will generate a unique ID, run the data frame through
the FFC, and then delete the results for that ID from the website so as not
to clutter up the user's account, or store too much data on the server side.
}
