% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{get_ffc_results_for_df}
\alias{get_ffc_results_for_df}
\title{Run Data Frame Through Functional Flows Calculator}
\usage{
get_ffc_results_for_df(flows_df, comid, flow_field, date_field, start_date)
}
\arguments{
\item{flows_df}{DataFrame. A time series data frame with flow and date columns}

\item{comid}{character. The COMID of the stream segment}

\item{flow_field}{character, default "flow". The name of the field in \code{df} that contains
flow values.}

\item{date_field}{character, default "date". The name of the field in \code{df} that contains
date values for each flow. The date field must be in MM/DD/YYYY format
as either factor or character values - true dates likely will not work
based on the API we're using. If you need to convert date values, add
a field to your existing data frame with the values in MM/DD/YYYY format
before providing it to this function.}

\item{start_date}{character, default "10/1". What month and day should the water
year start on? Neither month nor day needs to be zero-padded here, so
March first could just be 3/1, while December 12th can be 12/12.}
}
\value{
list of results from the functional flows calculator. More information will be
        forthcoming as we inspect the structure of what is returned.
}
\description{
This is primarily an internal function used to run data through the functional flows
calculator online, but is also available for those that wish to run the data themselves
and then do any other handling and transformation for postprocessing on their own.
}
\details{
Most people will want to use \code{\link{evaluate_alteration}} (for timeseries dataframes)
or \code{\link{evaluate_gage_alteration}} (for USGS gages) instead.

Internally, this is the primary function to use from the API client itself to obtain
raw FFC results. It will generate a unique ID, run the data frame through
the FFC, and then delete the results for that ID from the website so as not
to clutter up the user's account, or store too much data on the server side.
}
