% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{get_results_for_name}
\alias{get_results_for_name}
\title{Gets the results for the given named run of the FFC. Returns the nested list - no other processing}
\usage{
get_results_for_name(name, autodelete)
}
\description{
Gets the results for the given named run of the FFC. Returns the nested list - no other processing
}
