% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_api_client.R
\name{set_token}
\alias{set_token}
\title{Provide the token string used for accessing the Eflows site.
This is a function so that it can set the package private TOKEN variable}
\usage{
set_token(token_string)
}
\description{
Provide the token string used for accessing the Eflows site.
This is a function so that it can set the package private TOKEN variable
}
