% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteration.R
\name{single_metric_alteration}
\alias{single_metric_alteration}
\title{Assess the alteration of a single flow metric}
\usage{
single_metric_alteration(
  percentiles,
  predictions,
  ffc_values,
  days_in_water_year,
  annual
)
}
\arguments{
\item{percentiles}{data frame row - should have a named value "p50" that can be accessed, at the very least and a column
Metric with the flow metric in it. These are calculated percentile values from the FFC.}

\item{predictions}{data frame (or other named field item) the predicted flow metric values for the segment and metric}

\item{ffc_values}{vector of raw observed metric values (FFC output) for this metric}

\item{days_in_water_year}{numeric of how many days in the water year (defaults to 365, but could be 366).}
}
\description{
Given a metric's calculated percentiles, raw FFC output values, and predictions, returns a row of information indicating
whether or not that metric is likely altered, indeterminate, or likely unaltered. Includes fields with a text status,
an integer code (1=likely unaltered, 2=indeterminate, 3=likely altered), as well as for which direction alteration is (or may be)
in if it's indeterminate or likely altered (values are low/high or early/late for timing metrics)
}
