% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gages.R
\name{USGSGage}
\alias{USGSGage}
\title{USGS Gage Retrieval Tools}
\description{
This class retrieves data for a USGS gage.
}
\details{
#library(ffcAPIClient)
#gageid <- 11427000
#gage <- USGSGage$new()
#gage$id <- gageid
#gage$get_data()
#gage$get_comid()
#gage$comid
[1] 14996611
#ffcAPIClient::get_predicted_flow_metrics(gage$comid)
                Metric    COMID          p10          p25        p50          p75          p90 source
70804        DS_Dur_WS 14996611 1.051875e+02 1.273438e+02   154.0625 1.785563e+02 1.953908e+02  model
211050       DS_Mag_50 14996611 4.998793e+01 6.732828e+01   104.4028 1.464183e+02 1.882733e+02  model
351296       DS_Mag_90 14996611 9.314097e+01 1.291930e+02   173.6844 2.382053e+02 3.393799e+02  model
491542          DS_Tim 14996611 2.720000e+02 2.823875e+02   296.8875 3.070000e+02 3.210167e+02  model
586665          FA_Dur 14996611 2.000000e+00 3.000000e+00     4.0000 6.000000e+00 8.000000e+00    obs
702508          FA_Mag 14996611 1.294269e+02 1.886283e+02   289.6838 4.540329e+02 8.514823e+02  model
842754          FA_Tim 14996611 7.816667e+00 1.400000e+01    24.6250 2.900000e+01 4.217000e+01  model
983000         Peak_10 14996611 1.243107e+04 1.947545e+04 22830.3355 3.124928e+04 3.767889e+04  model
1123246        Peak_20 14996611 8.078893e+03 1.227363e+04 20218.4829 2.087196e+04 2.087196e+04  model
1263492        Peak_50 14996611 3.532988e+03 7.350986e+03  8542.1191 8.969386e+03 8.969386e+03  model
1358615    Peak_Dur_10 14996611 1.000000e+00 1.000000e+00     1.0000 2.000000e+00 4.000000e+00    obs
1429335    Peak_Dur_20 14996611 1.000000e+00 1.000000e+00     2.0000 3.000000e+00 6.000000e+00    obs
1500055    Peak_Dur_50 14996611 1.000000e+00 1.000000e+00     4.0000 1.000000e+01 2.900000e+01    obs
1570775    Peak_Fre_10 14996611 1.000000e+00 1.000000e+00     1.0000 1.000000e+00 2.000000e+00    obs
1641495    Peak_Fre_20 14996611 1.000000e+00 1.000000e+00     1.0000 2.000000e+00 3.000000e+00    obs
1712215    Peak_Fre_50 14996611 1.000000e+00 1.000000e+00     2.0000 3.000000e+00 5.000000e+00    obs
1828058         SP_Dur 14996611 4.700000e+01 5.900000e+01    72.0000 9.527500e+01 1.215417e+02  model
1968304         SP_Mag 14996611 1.067727e+03 1.662598e+03  2489.0563 3.771512e+03 5.809320e+03  model
2063427         SP_ROC 14996611 3.845705e-02 4.863343e-02     0.0625 8.132020e-02 1.141117e-01    obs
2179270         SP_Tim 14996611 1.607717e+02 1.905000e+02   218.7500 2.354750e+02 2.447583e+02  model
2319516    Wet_BFL_Dur 14996611 7.633333e+01 1.073000e+02   141.1958 1.633750e+02 1.875000e+02  model
2459762 Wet_BFL_Mag_10 14996611 1.519943e+02 1.960031e+02   278.2581 4.384614e+02 5.489183e+02  model
2600008 Wet_BFL_Mag_50 14996611 4.148992e+02 5.902507e+02   924.1728 1.175461e+03 1.426576e+03  model
2740254        Wet_Tim 14996611 4.937500e+01 5.905000e+01    73.0000 8.835625e+01 1.035083e+02  model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-validate}{\code{USGSGage$validate()}}
\item \href{#method-get_data}{\code{USGSGage$get_data()}}
\item \href{#method-get_comid}{\code{USGSGage$get_comid()}}
\item \href{#method-get_predicted_metrics}{\code{USGSGage$get_predicted_metrics()}}
\item \href{#method-clone}{\code{USGSGage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USGSGage$validate(latlong)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USGSGage$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_comid"></a>}}
\subsection{Method \code{get_comid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USGSGage$get_comid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_predicted_metrics"></a>}}
\subsection{Method \code{get_predicted_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USGSGage$get_predicted_metrics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USGSGage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
